Remove symlinks
Arduino Library doesn't want repos with symlinks, so we need to remove them.

We used symlinks because:
The Arduino library requires us to have our source in a directory called 'src'
in the root of the repo. However, the Arduino-Makefile looks for each library in a directory with its name.

To work around we tell ArduinoMakefile that our library oatmeal-protocol is
called 'src'.