List pyserial in install_requires (#10)
I thought at first that maybe it made sense to somehow reference this from our existing requirements.txt, but https://stackoverflow.com/a/33685899/1709587 makes a decent case that actually it's reasonable to duplicate these requirements and specify versions in the requirements.txt but not in the setup.py. So this solution seems fine.